+++ src/__test__/components/ContentWrapper.test.jsx
@@ -14,7 +14,9 @@ import AppRouteProvider from 'utils/AppRouteProvider';
 
 import { makeStore } from 'redux/store';
 import { getBackendStatus } from 'redux/selectors';
-import { loadProjects, setActiveProject } from 'redux/actions/projects';
+
+import { loadExperiments } from 'redux/actions/experiments';
+import { setActiveProject } from 'redux/actions/projects';
 import { updateExperimentInfo } from 'redux/actions/experimentSettings';
 import generateGem2sParamsHash from 'utils/data-management/generateGem2sParamsHash';
 
@@ -117,7 +119,7 @@ describe('ContentWrapper', () => {
       status: null,
     }));
 
-    await store.dispatch(loadProjects());
+    await store.dispatch(loadExperiments());
     await store.dispatch(setActiveProject(experimentId));
     await store.dispatch(updateExperimentInfo({ experimentId, experimentName, sampleIds }));
   });
