+++ src/__test__/components/data-management/SamplesTable.test.jsx
@@ -18,10 +18,10 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import createTestComponentFactory from '__test__/test-utils/testComponentFactory';
 
-import { loadProjects, setActiveProject } from 'redux/actions/projects';
+import { loadExperiments, loadExperiment } from 'redux/actions/experiments';
+import { setActiveProject } from 'redux/actions/projects';
 
 import loadEnvironment from 'redux/actions/networkResources/loadEnvironment';
-import { loadExperiments } from 'redux/actions/experiments';
 import { loadSamples } from 'redux/actions/samples';
 
 import mockDemoExperiments from '__test__/test-utils/mockData/mockDemoExperiments.json';
@@ -98,11 +98,11 @@ describe('Samples table', () => {
 
     storeState = makeStore();
 
-    await storeState.dispatch(loadProjects());
+    await storeState.dispatch(loadExperiments());
 
     // Loading experiment is usually called in Data Management, so we have to load them manually
-    await storeState.dispatch(loadExperiments(experimentWithSamplesId));
-    await storeState.dispatch(loadExperiments(experimentWithoutSamplesId));
+    await storeState.dispatch(loadExperiment(experimentWithSamplesId));
+    await storeState.dispatch(loadExperiment(experimentWithoutSamplesId));
     await storeState.dispatch(loadSamples(experimentWithSamplesId));
 
     // Defaults to project with samples
