+++ src/__test__/redux/actions/experiments/switchExperiment.test.js
@@ -1,14 +1,15 @@
-import { loadExperiments, switchExperiment } from 'redux/actions/experiments';
 import _ from 'lodash';
+import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
+
+import { makeStore } from 'redux/store';
+import { setActiveProject } from 'redux/actions/projects';
+import { loadExperiments, loadExperiment, switchExperiment } from 'redux/actions/experiments';
+import { loadSamples } from 'redux/actions/samples';
+
+import { responseData } from '__test__/test-utils/mockData';
 import mockAPI, {
   generateDefaultMockAPIResponses,
-
 } from '__test__/test-utils/mockAPI';
-import { makeStore } from 'redux/store';
-import { loadProjects, setActiveProject } from 'redux/actions/projects';
-import { responseData } from '__test__/test-utils/mockData';
-import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
-import { loadSamples } from 'redux/actions/samples';
 
 let store = null;
 
@@ -36,10 +37,10 @@ describe('switch experiment ', () => {
     fetchMock.mockIf(/.*/, mockAPI(mockAPIResponses));
 
     store = makeStore();
-    await store.dispatch(loadProjects());
-    await store.dispatch(loadExperiments(experimentWithSamplesId));
+    await store.dispatch(loadExperiments());
+    await store.dispatch(loadExperiment(experimentWithSamplesId));
     await store.dispatch(loadSamples(experimentWithSamplesId));
-    await store.dispatch(loadExperiments(experimentWithoutSamplesId));
+    await store.dispatch(loadExperiment(experimentWithoutSamplesId));
     await store.dispatch(setActiveProject(experimentWithoutSamplesId));
   });
