+++ src/__test__/redux/actions/projects/loadExperiments.test.js
@@ -3,7 +3,7 @@ import thunk from 'redux-thunk';
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
 import { PROJECTS_LOADED, PROJECTS_ERROR } from 'redux/actionTypes/projects';
 import { projectTemplate } from 'redux/reducers/projects/initialState';
-import { loadProjects } from 'redux/actions/projects';
+import { loadExperiments } from 'redux/actions/experiments';
 
 import fake from '__test__/test-utils/constants';
 import mockAPI, { generateDefaultMockAPIResponses } from '__test__/test-utils/mockAPI';
@@ -31,7 +31,7 @@ describe('load projects ', () => {
   it('Works correctly', async () => {
     const store = mockStore(initialState);
 
-    await store.dispatch(loadProjects());
+    await store.dispatch(loadExperiments());
 
     const actions = store.getActions();
 
@@ -43,7 +43,7 @@ describe('load projects ', () => {
   it('Dispatches error correctly', async () => {
     fetchMock.mockReject(new Error('Something went wrong :/'));
     const store = mockStore(initialState);
-    await store.dispatch(loadProjects());
+    await store.dispatch(loadExperiments());
     const action = store.getActions()[1];
     expect(action.type).toEqual(PROJECTS_ERROR);
   });
