+++ src/redux/actions/experiments/index.js
@@ -1,6 +1,7 @@
 import createExperiment from './createExperiment';
 import updateExperiment from './updateExperiment';
 import loadExperiments from './loadExperiments';
+import loadExperiment from './loadExperiment';
 import switchExperiment from './switchExperiment';
 import reorderSamples from './reorderSamples';
 
@@ -8,6 +9,7 @@ export {
   createExperiment,
   updateExperiment,
   loadExperiments,
+  loadExperiment,
   switchExperiment,
   reorderSamples,
 };
