+++ src/redux/actions/experiments/loadExperiments.js
@@ -1,73 +1,25 @@
 import fetchAPI from 'utils/http/fetchAPI';
 import handleError from 'utils/http/handleError';
+import endUserMessages from 'utils/endUserMessages';
 
-import {
-  // EXPERIMENTS_LOADED,
-  EXPERIMENTS_ERROR,
-  // EXPERIMENTS_LOADING,
-} from 'redux/actionTypes/experiments';
+import { EXPERIMENTS_ERROR, EXPERIMENTS_LOADED, EXPERIMENTS_LOADING } from 'redux/actionTypes/experiments';
 
-const toApiV1 = (experimentV2) => {
-  const {
-    id,
-    name,
-    description,
-    samplesOrder,
-    notifyByEmail,
-    processingConfig,
-    createdAt,
-    pipelines,
-    metadataKeys,
-  } = experimentV2;
-
-  const experimentV1 = {
-    experimentId: id,
-    projectId: id,
-    description,
-    experimentName: name,
-    processingConfig,
-    createdDate: createdAt,
-    notifyByEmail,
-    sampleIds: samplesOrder,
-    metadataKeys,
-    meta: {
-      // This is always 10x and organism so we can just generate them here
-      organism: null,
-      type: '10x',
-      gem2s: pipelines.gem2s ?? undefined,
-      pipeline: pipelines.qc ?? undefined,
-    },
-  };
-
-  return experimentV1;
-};
-
-const loadExperiments = (
-  projectUuid,
-) => async (dispatch) => {
-  // dispatch({
-  //   type: EXPERIMENTS_LOADING,
-  // });
+const loadExperiments = () => async (dispatch) => {
+  dispatch({
+    type: EXPERIMENTS_LOADING,
+  });
 
   try {
-    let data = await fetchAPI(`/v2/experiments/${projectUuid}`);
+    const data = await fetchAPI('/v2/experiments');
 
-    console.log('dataDebgu1');
-    console.log(data);
-
-    data = [toApiV1(data)];
-
-    console.log('dataDebug');
-    console.log(data);
-
-    // dispatch({
-    //   type: EXPERIMENTS_LOADED,
-    //   payload: {
-    //     experiments: data,
-    //   },
-    // });
+    dispatch({
+      type: EXPERIMENTS_LOADED,
+      payload: {
+        experiments: data,
+      },
+    });
   } catch (e) {
-    const errorMessage = handleError(e);
+    const errorMessage = handleError(e, endUserMessages.ERROR_LOADING_PROJECT);
 
     dispatch({
       type: EXPERIMENTS_ERROR,
@@ -77,5 +29,4 @@ const loadExperiments = (
     });
   }
 };
-
 export default loadExperiments;
